Added Browserify, Created QR Scanner
Creating a QR scanner required some restructuring of the project as before I was using a CDN for many elements, but now decided to use node modules, which meant that the code needed to work with 'require' statements, which is not support in the browser. So I added Browserify to bundle the files before linking them to the webpage. After that long process, it was a simple job of adding the qrcode library in and linking it to the camera. The scanner finally works!